Refactor `BayesianFunctionalRegression`
